[SQL] Move SLT into a separate project
