+++ sql-to-dbsp-compiler/SQL-compiler/pom.xml
@@ -4,9 +4,14 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>org.example</groupId>
+    <groupId>com.feldera</groupId>
     <artifactId>SQL-compiler</artifactId>
     <version>1.0-SNAPSHOT</version>
+    <parent>
+        <artifactId>dbsp-compiler</artifactId>
+        <groupId>com.feldera</groupId>
+        <version>1.0-SNAPSHOT</version>
+    </parent>
 
     <properties>
         <maven.compiler.source>8</maven.compiler.source>
@@ -186,7 +191,6 @@
             <artifactId>slf4j-simple</artifactId>
             <version>2.0.5</version>
         </dependency>
-        <!-- Keep the versions of the next two JARs in sync -->
         <dependency>
             <groupId>org.apache.calcite</groupId>
             <artifactId>calcite-core</artifactId>
@@ -212,13 +216,6 @@
             <artifactId>junit</artifactId>
             <version>4.13.2</version>
             <scope>test</scope>
-        </dependency>
-        <!-- HyperSQL Java database -->
-        <dependency>
-            <groupId>org.hsqldb</groupId>
-            <artifactId>hsqldb</artifactId>
-            <version>2.7.1</version>
-            <classifier>jdk8</classifier>
         </dependency>
         <!-- The following is necessary only if you want to use the JDBC Sqlite driver
              for testing -->
