+++ sql-to-dbsp-compiler/slt/src/main/java/org/dbsp/sqllogictest/SqlTestPrepareInput.java
