+++ sql-to-dbsp-compiler/slt/src/main/java/org/dbsp/sqllogictest/executors/InputFunctionGenerator.java
@@ -13,6 +13,7 @@ import org.dbsp.sqlCompiler.ir.statement.DBSPLetStatement;
 import org.dbsp.sqlCompiler.ir.statement.DBSPStatement;
 import org.dbsp.sqlCompiler.ir.type.*;
 import org.dbsp.sqlCompiler.ir.type.primitive.DBSPTypeVoid;
+import org.dbsp.sqllogictest.Main;
 import org.dbsp.util.Linq;
 import org.dbsp.util.TableValue;
 
@@ -83,7 +84,7 @@ class InputFunctionGenerator {
             if (this.connectionString.equals("csv")) {
                 // If the data is large write, it to a set of CSV files and read it at runtime.
                 for (int i = 0; i < inputSets.length; i++) {
-                    String fileName = (DBSPExecutor.rustDirectory + inputSets[i].tableName) + ".csv";
+                    String fileName = (Main.rustDirectory + inputSets[i].tableName) + ".csv";
                     File file = new File(fileName);
                     ToCsvVisitor.toCsv(compiler, file, inputSets[i].contents);
                     fields[i] = new DBSPApplyExpression(CalciteObject.EMPTY, "read_csv",
